Move project into right folder
