+++ .github/workflows/ci.yaml
@@ -1,0 +1,19 @@
+name: Invoker
+
+on:
+  pull_request:
+    types: [closed]
+
+jobs:
+  trigger-workflow:
+    runs-on: ubuntu-latest
+    if: github.event.pull_request.merged == true && github.event.pull_request.base.ref == 'main'
+    steps:
+      - name: Invoke workflow
+        uses: benc-uk/workflow-dispatch@v1
+        with:
+          workflow: ${{ vars.GH_BUILD_WORKFLOW }}
+          repo: ${{ vars.GH_BUILD_REPOSITORY }}
+          token: ${{ secrets.GH_PAT_TOKEN }}
+          ref: main
+          inputs: '{"repository_url": "${{ github.repository }}","ref":"${{ github.ref }}", "pullrequest_id":"${{ github.event.pull_request.number }}"}'
