+++ Apps.Hubspot.sln
@@ -3,14 +3,12 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.4.33213.308
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apps.Hubspot", "Apps.Hubspot\Apps.Hubspot\Apps.Hubspot.csproj", "{483D78DA-0409-4D48-BE76-6933F2AC8793}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apps.Hubspot", "Apps.Hubspot\Apps.Hubspot.csproj", "{483D78DA-0409-4D48-BE76-6933F2AC8793}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{12747CD6-147D-4E7D-9F74-91E18A89BBB0}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apps.Hubspot.Test", "Apps.Hubspot.Test\Apps.Hubspot.Test.csproj", "{AC854578-32F9-43E5-A28E-EBE5D4E2ECAA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
