+++ Apps.Hubspot/ApplicationConstants.cs
@@ -1,0 +1,17 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Apps.Hubspot
+{
+    public class ApplicationConstants
+    {
+        public const string ClientId = "#{HUBSPOTCMS_CLIENT_ID}";
+
+        public const string Scope = "#{HUBSPOTCMS_SCOPE}";
+
+        public const string ClientSecret = "#{HUBSPOTCMS_SECRET}";
+    }
+}
