+++ Apps.Hubspot/Auth/OAuth2/OAuth2AuthorizeService.cs
@@ -1,0 +1,31 @@
+﻿using Apps.Hubspot.Constants;
+using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Authentication.OAuth2;
+using Blackbird.Applications.Sdk.Common.Invocation;
+using Blackbird.Applications.Sdk.Utils.Extensions.String;
+using Microsoft.AspNetCore.WebUtilities;
+
+namespace Apps.Hubspot.Auth.OAuth2;
+
+public class OAuth2AuthorizeService : BaseInvocable, IOAuth2AuthorizeService
+{
+    public OAuth2AuthorizeService(InvocationContext invocationContext) : base(invocationContext)
+    {
+    }
+
+    public string GetAuthorizationUrl(Dictionary<string, string> values)
+    {
+        string bridgeOauthUrl = $"{InvocationContext.UriInfo.BridgeServiceUrl.ToString().TrimEnd('/')}/oauth";
+        var parameters = new Dictionary<string, string>
+        {
+            { "client_id", ApplicationConstants.ClientId },
+            { "redirect_uri", $"{InvocationContext.UriInfo.BridgeServiceUrl.ToString().TrimEnd('/')}/AuthorizationCode" },
+            { "scope", ApplicationConstants.Scope },
+            { "state", values["state"] },
+            { "authorization_url", Urls.OAuth},
+            { "actual_redirect_uri", InvocationContext.UriInfo.AuthorizationCodeRedirectUri.ToString() },
+        };
+
+        return QueryHelpers.AddQueryString(bridgeOauthUrl, parameters);
+    }
+}
