+++ Apps.Hubspot/Auth/OAuth2/OAuth2TokenService.cs
@@ -1,11 +1,17 @@
 ﻿using Apps.Hubspot.Constants;
+using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Authentication.OAuth2;
+using Blackbird.Applications.Sdk.Common.Invocation;
 using Newtonsoft.Json;
 
 namespace Apps.Hubspot.Auth.OAuth2;
 
-public class OAuth2TokenService : IOAuth2TokenService
+public class OAuth2TokenService : BaseInvocable, IOAuth2TokenService
 {
+    public OAuth2TokenService(InvocationContext invocationContext) : base(invocationContext)
+    {
+    }
+
     public bool IsRefreshToken(Dictionary<string, string> values)
     {
         var expiresAt = DateTime.Parse(values[CredsNames.ExpiresAt]);
@@ -36,7 +42,7 @@ public class OAuth2TokenService : IOAuth2TokenService
             { "grant_type", "authorization_code" },
             { "client_id", ApplicationConstants.ClientId },
             { "client_secret", ApplicationConstants.ClientSecret },
-            { "redirect_uri", ApplicationConstants.RedirectUri },
+            { "redirect_uri", $"{InvocationContext.UriInfo.BridgeServiceUrl.ToString().TrimEnd('/')}/AuthorizationCode" },
             { "code", code }
         };
