+++ Apps.Hubspot/Connections/ConnectionValidator.cs
