+++ Apps.Hubspot/Connections/OAuth2ConnectionDefinition.cs
