+++ Apps.Hubspot/Exceptions/HubspotException.cs
