+++ Apps.Hubspot/Helpers/PageHelpers.cs
