+++ Apps.Hubspot/Models/Requests/BlogPosts/BlogPostRequest.cs
