+++ Apps.Hubspot/Models/Requests/BlogPosts/GetBlogPostTranslationRequest.cs
