+++ Apps.Hubspot/Models/Requests/BlogPosts/ManageBlogPostRequest.cs
