+++ Apps.Hubspot/Models/Requests/BlogPosts/TranslateBlogPostFromHtmlRequest.cs
