+++ Apps.Hubspot/Models/Requests/Translations/CreateTranslationRequest.cs
