+++ Apps.Hubspot/Models/Responses/Error.cs
