+++ Apps.Hubspot/Models/Responses/Files/FileResponse.cs
