+++ Apps.Hubspot/Models/Responses/GetAllResponse.cs
