+++ Apps.Hubspot/Models/Responses/Pages/PageInfoResponse.cs
