+++ Apps.Hubspot/Models/Responses/Translations/TranslationResponse.cs
